refactor(Controller, View): Contoller 생성 시점에 뷰 주입받도록 수정
