+++ src/main/kotlin/Controller.kt
@@ -3,36 +3,41 @@ import domain.OmokBoard
 import domain.OmokGame
 import domain.State
 import domain.Stone
+import view.ConsoleInputView
+import view.ConsoleOutputView
 import view.InputView
 import view.OutputView
 
-class Controller {
+class Controller(
+    private val inputView: InputView = ConsoleInputView(),
+    private val outputView: OutputView = ConsoleOutputView()
+) {
 
     private val omokGameListener = object : Listener {
         override fun onStoneRequest(): Stone {
-            return InputView.readPosition()
+            return inputView.readPosition()
         }
 
         override fun onMove(omokBoard: OmokBoard, state: State, stone: Stone) {
-            OutputView.printOmokState(omokBoard, state, stone)
+            outputView.printOmokState(omokBoard, state, stone)
         }
 
         override fun onMoveFail() {
-            OutputView.printDuplicate()
+            outputView.printDuplicate()
         }
 
         override fun onForbidden() {
-            OutputView.printForbidden()
+            outputView.printForbidden()
         }
 
         override fun onFinish(state: State) {
-            OutputView.printWinner(state)
+            outputView.printWinner(state)
         }
     }
 
     fun run() {
         val omokGame = OmokGame(omokGameListener = omokGameListener)
-        OutputView.printStart()
+        outputView.printStart()
         omokGame.runGame()
     }
 }
